Move exchanges code into their own directory
