+++ rotkehlchen/exchanges/__init__.py
@@ -1,0 +1,5 @@
+from rotkehlchen.exchanges.binance import Binance  # noqa: F401
+from rotkehlchen.exchanges.bitmex import Bitmex  # noqa: F401
+from rotkehlchen.exchanges.bittrex import Bittrex  # noqa: F401
+from rotkehlchen.exchanges.kraken import Kraken  # noqa: F401
+from rotkehlchen.exchanges.poloniex import Poloniex  # noqa: F401
