+++ rotkehlchen/history.py
@@ -4,9 +4,6 @@ from json.decoder import JSONDecodeError
 from typing import TYPE_CHECKING, List, Optional, Tuple, Union
 
 from rotkehlchen.assets.asset import Asset
-from rotkehlchen.binance import Binance
-from rotkehlchen.bitmex import Bitmex, trade_from_bitmex
-from rotkehlchen.bittrex import Bittrex
 from rotkehlchen.constants.assets import A_BTC
 from rotkehlchen.db.dbhandler import DBHandler
 from rotkehlchen.errors import (
@@ -16,10 +13,12 @@ from rotkehlchen.errors import (
     UnknownAsset,
     UnsupportedAsset,
 )
-from rotkehlchen.exchange import data_up_todate
+from rotkehlchen.exchanges import Binance, Bitmex, Bittrex, Kraken, Poloniex
+from rotkehlchen.exchanges.bitmex import trade_from_bitmex
+from rotkehlchen.exchanges.exchange import data_up_todate
+from rotkehlchen.exchanges.poloniex import process_polo_loans, trade_from_poloniex
 from rotkehlchen.fval import FVal
 from rotkehlchen.inquirer import Inquirer
-from rotkehlchen.kraken import Kraken
 from rotkehlchen.logging import RotkehlchenLogsAdapter
 from rotkehlchen.order_formatting import (
     AssetMovement,
@@ -29,7 +28,6 @@ from rotkehlchen.order_formatting import (
     asset_movements_from_dictlist,
     trades_from_dictlist,
 )
-from rotkehlchen.poloniex import Poloniex, process_polo_loans, trade_from_poloniex
 from rotkehlchen.transactions import query_etherscan_for_transactions, transactions_from_dictlist
 from rotkehlchen.typing import EthAddress, FiatAsset, FilePath, Price, Timestamp
 from rotkehlchen.user_messages import MessagesAggregator
