+++ rotkehlchen/tests/utils/accounting.py
@@ -1,7 +1,7 @@
 from typing import Any, Dict, List
 
+from rotkehlchen.exchanges.poloniex import process_polo_loans
 from rotkehlchen.order_formatting import asset_movements_from_dictlist, trades_from_dictlist
-from rotkehlchen.poloniex import process_polo_loans
 from rotkehlchen.transactions import transactions_from_dictlist
 from rotkehlchen.typing import Timestamp
